Removes warnings in several modules when built on linux
This appears to speed up compilation time with clang and produces much
cleaner compiler output. Problematic warnings targeted by these changes
include -Wswitch, -Wdelete-incomplete, -Wconstant-logical-operand,
-Wshift-negative-value, and -Winconsistent-missing-override.

Builds successfully on linux and clang 3.9.1 with
ENABLE_MODULE_IRRLICHT, ENABLE_MODULE_PARALLEL, ENABLE_MODULE_VEHICLE,
and -DCMAKE_CXX_FLAGS=-Werror in Debug and Release configurations.

Also builds successfully on linux and GCC 6.2.1 with the same
configuration options.

+++ src/chrono/collision/bullet/BulletCollision/CollisionShapes/btCompoundShape.cpp
@@ -78,7 +78,7 @@ void	btCompoundShape::addChildShape(const btTransform& localTransform,btCollisio
 	{
 		const btDbvtVolume	bounds=btDbvtVolume::FromMM(localAabbMin,localAabbMax);
 		int index = m_children.size();
-		child.m_node = m_dynamicAabbTree->insert(bounds,(void*)index);
+		child.m_node = m_dynamicAabbTree->insert(bounds,(void*)(long)index);
 	}
 
 	m_children.push_back(child);
@@ -307,7 +307,7 @@ void btCompoundShape::createAabbTreeFromChildren()
             child.m_childShape->getAabb(child.m_transform,localAabbMin,localAabbMax);
 
             const btDbvtVolume  bounds=btDbvtVolume::FromMM(localAabbMin,localAabbMax);
-            child.m_node = m_dynamicAabbTree->insert(bounds,(void*)index);
+            child.m_node = m_dynamicAabbTree->insert(bounds,(void*)(long)index);
         }
     }
 }
